Revert unnecessary moving constant and render filtering to separate files.
